orderitem modificated

+++ src/main/java/com/example/unimagdalena/TiendaEcommerce/entities/OrderItem.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Entity
@@ -21,13 +23,13 @@ public class OrderItem {
         private Long id;
 
         @Column(nullable = false)
-        private Double unitPrice;
+        private BigDecimal unitPrice;
 
         @Column(nullable = false)
         private Integer quantity;
 
         @Column(nullable = false)
-        private Double subtotal;
+        private BigDecimal subtotal;
 
         @ManyToOne(fetch = FetchType.LAZY)
         @JoinColumn(name = "order_id", nullable = false)
